clean up rational.typeexpr
